rollback from 2.5 register flow, without loading animations, changed stackRouter -> switchRouter flow to home component after registration, fixing for many times firebase init
